Added the muscle groups pages and refactored slightly
