+++ krafted-physique/src/App.css
@@ -190,7 +190,7 @@ a {
 }
 
 /* topRated */
-.topRatedContainer {
+.pagesContainer {
     width: 100%;
     min-height: 100vh;
     align-items: center;
@@ -337,7 +337,7 @@ a {
     padding: 3rem 2rem;
 }
 
-.titleFAQ, .titleTopRated {
+.titleFAQ, .pagesTitle {
     border-bottom: solid 1px rgba(0, 0, 255, 0.701);
     padding-bottom: 1rem;
 }
