+++ krafted-physique/src/App.jsx
@@ -5,6 +5,11 @@ import Home from "./Components/Home"
 import TopRated from "./Components/TopRated"
 import FAQs from "./Components/FAQ.jsx"
 import ContactUs from "./Components/ContactUs"
+import Chest from "./Components/MusclePages/Chest.jsx"
+import Back from "./Components/MusclePages/Back.jsx"
+import Shoulders from "./Components/MusclePages/Shoulders.jsx"
+import Legs from "./Components/MusclePages/Legs.jsx"
+import Arms from "./Components/MusclePages/Arms.jsx"
 
 export default function App(){
 
@@ -13,6 +18,11 @@ export default function App(){
       <Navbar/>
       <Routes>
         <Route path="/" element={<Home />}/>
+        <Route path="/Chest" element={<Chest />}/>
+        <Route path="/Back" element={<Back />}/>
+        <Route path="/Shoulders" element={<Shoulders />}/>
+        <Route path="/Legs" element={<Legs />}/>
+        <Route path="/Arms" element={<Arms />}/>
         <Route path="/TopRated" element={<TopRated />}/>
         <Route path="/FAQs" element={<FAQs />}/>
         <Route path="/ContactUs" element={<ContactUs />}/>
