+++ krafted-physique/src/Components/MusclePages/Arms.jsx
@@ -1,5 +1,10 @@
+import PageTemplate from "../Template/PageTemplate"
+
 export default function Arms() {
-    return(
-        ''
+    return (
+        <PageTemplate 
+            url={'arms'}
+            title={'Chisel Your Arms: Proven Exercise Techniques'}
+        />
     )
 }
