+++ krafted-physique/src/Components/MusclePages/Back.jsx
@@ -1,5 +1,10 @@
+import PageTemplate from "../Template/PageTemplate"
+
 export default function Back() {
-    return(
-        ''
+    return (
+        <PageTemplate 
+            url={'back'}
+            title={'Back to Basics: Proven Methods for Toning and Strength'}
+        />
     )
 }
