+++ krafted-physique/src/Components/MusclePages/Shoulders.jsx
@@ -1,5 +1,10 @@
+import PageTemplate from "../Template/PageTemplate"
+
 export default function Shoulders() {
-    return(
-        ''
+    return (
+        <PageTemplate 
+            url={'shoulders'}
+            title={'Shoulder Sculpting: Enhance Strength and Definition'}
+        />
     )
 }
